Tools: change typing of get[] methods
The getPokemon(), getMove(), getItem(), getAbility(), and getFormat() methods may now return null. The main purpose of this change is to ensure that user input is being validated before use.

The getExisting[]() counterparts should only be used when you are not handling user input (such as looping through Tools.data) or it has already been validated and stored (such as in a game).